perf: stop blocking the host actor on configure
The app sends `configure` at startup, and answering it walked every session
in series through two two-second write barriers and a full table scan each,
on the single server actor. Every attach queued behind that.

- `trim_session` finds its cut with a window function instead of pulling
  `length(data)` for every row into Rust. Retention is by bytes, not rows, so
  a session persisting a batch every 100 ms accumulates six figures of them,
  and this runs on every checkpoint tick, every detach and every configure.
- `read_buffer` sizes its output in SQL rather than decoding every blob once
  for its length and again to copy it.
- `configure` applies the new cap to the in-memory buffers and hands the
  durable trim to a spawned task. The cap on disk is housekeeping; a write
  landing after the trim just leaves the session briefly over it, and the
  checkpoint tick trims again.

Left alone: a reminted session parses its restored scrollback for a title
twice, once in `restore_exited` and once in `Session::start`. Removing the
first would strand `restore_exited` as test-only code, which is a bigger
change than the one byte-loop is worth without a measurement to justify it.
